Restructure layout components
