+++ app/components/layout/predictive-search/predictive-search-results.tsx
