+++ app/components/layout/predictive-search/result-item.tsx
