+++ app/components/layout/scrolling-announcement.tsx
